Added info to the wiki (HTML) I tried to make the bullet and character move diagonally(i couldn't test it bc my SQLite db is locked) Changed variables from int to double for image movement

+++ src/renderer/Background.java
@@ -13,13 +13,13 @@ import javax.imageio.ImageIO;
 
 public class Background {
 
-	public int x;
-	public int y;
-	public int oX, oY;
+	public double x;
+	public double y;
+	public double oX, oY;
 	WorldLayout topDwn;
 	String fileName;
 
-	public Background(int x, int y, WorldLayout TopDwn, String fileName) {
+	public Background(double x, double y, WorldLayout TopDwn, String fileName) {
 		this.oX = x;
 		this.oY = y;
 		this.fileName = fileName;
@@ -36,8 +36,9 @@ public class Background {
 	}
 
 	public void render(Graphics g) throws IOException {
+		
 		BufferedImage manager = ImageIO.read(new File(this.fileName));
-		g.drawImage(manager, this.x, this.y, null);
+		g.drawImage(manager, (int) this.x,(int) this.y, null);
 
 	}
